Python: break tests into multiple files, add exhaustive tests to tox.
Mostly needed to find passenv in the tox configs.
