+++ python/tox.toml
@@ -3,7 +3,8 @@ env_list = ["format", "type", "3.13", "3.12", "3.11", "3.10", "3.9"]
 
 [env_run_base]
 description = "run unit tests"
-commands = [["python3", "-m", "unittest", { replace = "posargs", default = ["tests"], extend = true }]]
+commands = [["python3", "-m", "unittest", { replace = "posargs", default = ["tests/test_chr_ord.py", "tests/test_unit.py", "tests/test_integration.py", "tests/test_exhaustive.py"], extend = true }]]
+passenv = ["RUN_EXHAUSTIVE_TESTS"]
 
 [env.format]
 description = "run formatter"
